Attendant {waiter}	R Documentation

Attendant

Description

Manage the attendant loading bar with bootstrap 4.

Active bindings

max

Maximum value of the bar.

Methods

Public methods

Attendant$new()

Attendant$inc()

Attendant$dec()

Attendant$set()

Attendant$done()

Attendant$close()

Attendant$auto()

Attendant$getMin()

Attendant$getMax()

Attendant$getValue()

Attendant$clone()

Method new()

Usage

Attendant$new(
  id,
  min = NULL,
  max = NULL,
  session = shiny::getDefaultReactiveDomain(),
  hide_on_max = FALSE
)

Arguments

id

Id of progress bar set with [attendantBar].

min, max

Minimum and maximum value of the progress bar.

session

A valid shiny session.

hide_on_max

Whether to hide the progress bar when it reaches its maximum value (defined in [attendantBar]). The progress bar automatically becomes visible again when it is set to a value below the maximum.

Details

Initialise a progress bar

Method inc()

Usage

Attendant$inc(value = 1, text = NULL)

Arguments

value

Value to increase the progress bar.

text

Text to display on the progress bar.

Details

Increase

Method dec()

Usage

Attendant$dec(value = 1, text = NULL)

Arguments

value

Value to decrease the progress bar.

text

Text to display on the progress bar.

Details

Decrease

Method set()

Usage

Attendant$set(value, text = NULL)

Arguments

value

Value to set the progress bar.

text

Text to display on the progress bar.

Details

Set

Method done()

Usage

Attendant$done(text = NULL)

Arguments

text

Text to display on the progress bar.

Details

Done with progress

Method close()

Usage

Attendant$close(text = NULL)

Arguments

text

Text to display on the progress bar.

Details

Done with progress

Method auto()

Usage

Attendant$auto(ms = 400, value = 1)

Arguments

ms

Milliseconds between increment of 'value'.

value

Value to increment by at every 'ms'.

Details

Automatically increase the progress bar until done

Method getMin()

Usage

Attendant$getMin()

Details

Get minimum value

Method getMax()

Usage

Attendant$getMax()

Details

Get maximum value

Method getValue()

Usage

Attendant$getValue()

Details

Get current value

Method clone()

The objects of this class are cloneable with this method.

Usage

Attendant$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package waiter version 0.2.5.1 Index]