attendantBar {waiter}	R Documentation

Bootstrap 4 Loading Bar

Description

Create a Bootstrap 4 progress bar.

Usage

attendantBar(
  id,
  value = 0,
  min = 0,
  max = 100,
  text = NULL,
  color = c("primary", "info", "success", "danger", "warning"),
  striped = FALSE,
  animated = FALSE,
  height = 20,
  width = "100%",
  class = "",
  style = "",
  bg_color = "#f5f5f5",
  hidden = FALSE
)

Arguments

id
A unique identifier for the progress bar. Used in 'Attendant' class for handling.

value, min, max
Initial value, minimum, and maximum values the progress bar can take.

text
Optional text to display on the progress bar. This can then be dynamically modified with 'Attendant'.

striped
Whether the progress bar should be striped.

animated
Whether to animate the stripe on the progress bar.

height
Height of the progress bar, numerical values are converted to pixels ('px' CSS), any other valid CSS size is valid too.

width
Width of the bar, defaults to '100 values (e.g.: '42') are converted to pixels ('px').

class, style
Additional style and class to pass to the parent wrapper of the progress bar.

bg_color, color
Color, and background color of the progress bar.

hidden
Set to 'TRUE' to initialise the attendant as hidden, it will be made visible when set to a value.

[Package waiter version 0.2.5.1 Index]