spinners {waiter}	R Documentation

Spinners

Description

Spinkit spinners to use with waiter_show.

Usage

spin_rotating_plane()

spin_fading_circles()

spin_folding_cube()

spin_double_bounce()

spin_wave()

spin_wandering_cubes()

spin_pulse()

spin_chasing_dots()

spin_three_bounce()

spin_circle()

spin_rotate()

spin_solar()

spin_orbit()

spin_squares()

spin_cube_grid()

spin_circles()

spin_orbiter()

spin_pixel()

spin_flower()

spin_dual_ring()

spin_heart()

spin_ellipsis()

spin_facebook()

spin_hourglass()

spin_ring()

spin_ripple()

spin_terminal()

spin_loader()

spin_throbber()

spin_refresh()

spin_heartbeat()

spin_gauge()

spin_3k()

spin_wobblebar()

spin_atebits()

spin_whirly()

spin_flowers()

spin_dots()

spin_3circles()

spin_plus()

spin_pulsar()

spin_hexdots()

spin_inner_circles()

spin_pong()

spin_timer()

spin_ball()

spin_dual_circle()

spin_seven_circle()

spin_clock()

spin_pushing_shapes()

spin_fill()

spin_rhombus()

spin_balance()

spin_square_circle()

spin_circle_square()

spin_puzzle()

spin_half()

spin_loaders(id = 1, color = "white", style = NULL)

spin_1()

spin_2()

spin_3()

spin_4()

spin_5()

spin_6()

spin_google()

Arguments

id
The spinner identifier, an integer between 1, and 42.

color
Desired color of spinner.

style
CSS style to apply to spinner.

Details

Much of the CSS used is to provide those spinners. One can greatly reduce the load on the browser by only sourcing the CSS for the spinners required. You can find out which CSS kits are required to load by using the spinner in the R console as shown in the example. This prints the kit and instructions to only source the required file.

Value

An object of class spinner.

Examples

spin_rotating_plane()

[Package waiter version 0.2.0 Index]