waiterTheme {waiter}	R Documentation

Define a Theme

Description

Define a theme to be used by all waiter loading screens. These can be overriden in individual loading screens.

Usage

waiter_set_theme(html = spin_1(), color = "#333e48", logo = "", image = "")

waiter_get_theme()

waiter_unset_theme()

Arguments

html
HTML content of waiter, generally a spinner, see spinners.

color
Background color of loading screen.

logo
Path to logo to display. Deprecated.

image
Path to background image.

[Package waiter version 0.2.5.1 Index]