withProgressAttendant {waiter}	R Documentation

Report Progress Attendant

Description

Report progress with attendant.

Usage

withProgressAttendant(
  expr,
  ...,
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE
)

setProgressAttendant(
  value = 1,
  text = NULL,
  session = getDefaultReactiveDomain()
)

incProgressAttendant(
  value = 1,
  text = NULL,
  session = getDefaultReactiveDomain()
)

Arguments

expr
The work to be done. This expression should contain calls to setProgressAttendant' or 'incProgressAttendant'.

...
Passed to the Attendant constructor ('Attendant$new()').

session
The Shiny session object, as provided by 'shinyServer' to the server function. The default is to automatically find the session by using the current reactive domain.

env
The environment in which 'expr' should be evaluated.

quoted
Whether 'expr' is a quoted expression (this is not common).

value
Value to set the waitress to or increase it by.

text
Text to display on the progress bar.

[Package waiter version 0.2.5.1 Index]