withProgressWaitress {waiter}	R Documentation

Report Progress Waitress

Description

Report progress with waitress.

Usage

withProgressWaitress(
  expr,
  ...,
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE
)

setProgressWaitress(value = 1, session = getDefaultReactiveDomain())

incProgressWaitress(value = 1, session = getDefaultReactiveDomain())

Arguments

expr
The work to be done. This expression should contain calls to setProgressWaitress' or 'incProgressWaitress'.

...
Passed to the Waitress constructor ('Waitress$new()').

session
The Shiny session object, as provided by 'shinyServer' to the server function. The default is to automatically find the session by using the current reactive domain.

env
The environment in which 'expr' should be evaluated.

quoted
Whether 'expr' is a quoted expression (this is not common).

value
Value to set the waitress to or increase it by.

[Package waiter version 0.2.5.1 Index]