withWaiter {waiter}	R Documentation

With Waiter

Description

Adds a waiter to a rective UI element. Thew waiter is displayed when the element is invalidated then is removed when the element receives a new value.

Usage

withWaiter(element, html = spin_1(), color = "#333e48", image = "")

Arguments

element
A reactive element, e.g.: 'uiOutput', or 'plotOutput'.

html
HTML content of waiter, generally a spinner, see spinners.

color
Background color of loading screen.

image
Path to background image.

[Package waiter version 0.2.5.1 Index]