waiterTheme {waiter}	R Documentation

Define a Theme

Description

Define a theme to be used by all waiter loading screens. These can be overriden in individual loading screens.

Usage

waiter_set_theme(html = spin_1(), color = "#333e48", logo = "")

waiter_get_theme()

waiter_unset_theme()

Arguments

html
HTML content of waiter, generally a spinner, see spinners.

color
Background color of loading screen.

logo
Logo to display.

[Package waiter version 0.2.0 Index]